Fixed crashes with maven

+++ src/main/java/com/wmbest/jenesis/hardware/SixtyEightK.java
@@ -8,8 +8,8 @@ public class SixtyEightK {
     private UInt32 PC = new UInt32();
     private UInt32 SR = new UInt32();
 
-    private byte ccr() {
-        return ((byte) SR.get()) & 0x1fb;
+    private long ccr() {
+        return SR.get() & 0x1fb;
     }
 
     private Memory mMemory;
